Enhanced variable management from add-in
- Collapse tree on open
- Enhanced tree search
- Simplified variable creation

+++ ExcelAddIn/Classes/Types/Values3D.cs
@@ -25,7 +25,7 @@ namespace Avinet.Adaptive.Statistics.ExcelAddIn
             this[pRowIndex][pFieldIndex] = pValue;
         }
 
-        public int? GetValueByFieldRow(int pFieldIndex, int pRowIndex)
+        public double? GetValueByFieldRow(int pFieldIndex, int pRowIndex)
         {
             if (!this.ContainsKey(pRowIndex))
             {
